pyproject.toml
[tool.poetry]
name = "testfilefield"
version = "0.1.0"
description = ""
authors = ["long <longnguyen176251@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.5"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
